feat(PAYMENTS-21797): add form submit action
