--- src/features/headless-checkout/web-components/submit-button/default-submit-button/default-submit-button.handler.ts
@@ -1,14 +1,0 @@
-import { Handler } from '../../../../../core/post-messages-client/handler.type';
-import { Message } from '../../../../../core/message.interface';
-import { EventName } from '../../../../../core/event-name.enum';
-
-export const defaultSubmitButtonHandler: Handler<void> = (
-  message: Message
-): { isHandled: boolean } | null => {
-  if (message.name === EventName.submitForm) {
-    return {
-      isHandled: true,
-    };
-  }
-  return null;
-};
